requirements.txt
python-dotenv
openai
ipykernel
PyMuPDF